.pub-cache/hosted/pub.dartlang.org/html-0.15.1/test/data/tree-construction/tests14.dat

#data
<!DOCTYPE html><html><body><xyz:abc></xyz:abc>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <xyz:abc>

#data
<!DOCTYPE html><html><body><xyz:abc></xyz:abc><span></span>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <xyz:abc>
|     <span>

#data
<!DOCTYPE html><html><html abc:def=gh><xyz:abc></xyz:abc>
#errors
15: Unexpected start tag html
#document
| <!DOCTYPE html>
| <html>
|   abc:def="gh"
|   <head>
|   <body>
|     <xyz:abc>

#data
<!DOCTYPE html><html xml:lang=bar><html xml:lang=foo>
#errors
15: Unexpected start tag html
#document
| <!DOCTYPE html>
| <html>
|   xml:lang="bar"
|   <head>
|   <body>

#data
<!DOCTYPE html><html 123=456>
#errors
#document
| <!DOCTYPE html>
| <html>
|   123="456"
|   <head>
|   <body>

#data
<!DOCTYPE html><html 123=456><html 789=012>
#errors
#document
| <!DOCTYPE html>
| <html>
|   123="456"
|   789="012"
|   <head>
|   <body>

#data
<!DOCTYPE html><html><body 789=012>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     789="012"
